requirements.txt
# FastAPI and dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Async PostgreSQL driver
asyncpg==0.29.0

# SQLite for development
aiosqlite==0.19.0

# Environment variables
python-dotenv==1.0.0

# API clients
httpx==0.26.0
aiohttp==3.9.1

# Authentication
pyjwt==2.8.0
passlib[bcrypt]==1.7.4

# Google Sheets
google-api-python-client==2.114.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Image processing
pillow==10.2.0

# Background tasks
celery==5.3.6
redis==5.0.1

# CORS
fastapi-cors==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging
python-json-logger==2.0.7

# OpenAI
openai==1.9.0

# JWT for KLING
cryptography==41.0.7